Avoid defaulting to Dataflow type in parse_urn (possible silent mistyping)
