+++ tests/util/test_parse_any_urn.py
@@ -60,13 +60,8 @@ def test_match_short_item():
     assert m.item_id == "A"
 
 
-def test_match_short_dataflow():
+def test_short_dataflow_no_match():
     df = "BIS:CBS(1.0)"
 
-    m = parse_urn(df)
-
-    assert isinstance(m, Reference)
-    assert m.sdmx_type == "Dataflow"
-    assert m.agency == "BIS"
-    assert m.id == "CBS"
-    assert m.version == "1.0"
+    with pytest.raises(Invalid):
+        parse_urn(df)
